Update docs for new refernces section handling
Note that the need to distinguish between references-ness and not via
the title of the section has to do with disambiguating XSLT rules.
There are better ways to do this though, and we should.  One way might
be to match on the existence of bibxml inside the <hN> being matched, or
perhaps we'll want a new custom inset by which to mark which sections
are references sections.