Just a note about section depth handling
Should modify the lyx2xml.py script to add a section depth attribute
with a numeric value in all layout:*ection elements!  Might as well
rename them all "Section" while we're at it.  Of course, this will be
rather specific to document classes that have sections like this, so
lyx2xml won't necessarily be that generic anymore.  Something to think
about.

Or is there a simple way to derive section depth from the element name
in XPath?!  Well, we can always have an if/elif in XPath... but it's
unwieldy...